refactor: DAM-11 UseCase 패키지 구분
