+++ domain/src/main/java/com/khnsoft/damta/domain/di/UserUseCaseModule.kt
@@ -1,14 +1,14 @@
 package com.khnsoft.damta.domain.di
 
-import com.khnsoft.damta.domain.request.SignInRequest
-import com.khnsoft.damta.domain.request.SignUpRequest
-import com.khnsoft.damta.domain.request.UpdateUserRequest
+import com.khnsoft.damta.domain.request.user.SignInRequest
+import com.khnsoft.damta.domain.request.user.SignUpRequest
+import com.khnsoft.damta.domain.request.user.UpdateUserRequest
 import com.khnsoft.damta.domain.response.EmptyResponse
-import com.khnsoft.damta.domain.response.UpdateUserResponse
+import com.khnsoft.damta.domain.response.user.UpdateUserResponse
 import com.khnsoft.damta.domain.usecase.ResultUseCase
-import com.khnsoft.damta.domain.usecase.SignInUseCase
-import com.khnsoft.damta.domain.usecase.SignUpUseCase
-import com.khnsoft.damta.domain.usecase.UpdateUserUseCase
+import com.khnsoft.damta.domain.usecase.user.SignInUseCase
+import com.khnsoft.damta.domain.usecase.user.SignUpUseCase
+import com.khnsoft.damta.domain.usecase.user.UpdateUserUseCase
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
