--- domain/src/main/java/com/khnsoft/damta/domain/request/SignInRequest.kt
@@ -1,6 +1,0 @@
-package com.khnsoft.damta.domain.request
-
-data class SignInRequest(
-    val username: String,
-    val password: String
-) : Request
