+++ domain/src/main/java/com/khnsoft/damta/domain/request/user/SignInRequest.kt
@@ -1,0 +1,8 @@
+package com.khnsoft.damta.domain.request.user
+
+import com.khnsoft.damta.domain.request.Request
+
+data class SignInRequest(
+    val username: String,
+    val password: String
+) : Request
