+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/user/SignInUseCase.kt
@@ -1,10 +1,11 @@
-package com.khnsoft.damta.domain.usecase
+package com.khnsoft.damta.domain.usecase.user
 
 import com.khnsoft.damta.common.extension.flatMap
 import com.khnsoft.damta.domain.common.HashGenerator
 import com.khnsoft.damta.domain.repository.UserRepository
-import com.khnsoft.damta.domain.request.SignInRequest
+import com.khnsoft.damta.domain.request.user.SignInRequest
 import com.khnsoft.damta.domain.response.EmptyResponse
+import com.khnsoft.damta.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
 internal class SignInUseCase @Inject constructor(
