+++ domain/src/test/java/com/khnsoft/damta/domain/usecase/user/SignInUseCaseTest.kt
@@ -1,9 +1,9 @@
-package com.khnsoft.damta.domain.usecase
+package com.khnsoft.damta.domain.usecase.user
 
 import com.khnsoft.damta.domain.common.HashGenerator
 import com.khnsoft.damta.domain.error.UserError
 import com.khnsoft.damta.domain.repository.UserRepository
-import com.khnsoft.damta.domain.request.SignInRequest
+import com.khnsoft.damta.domain.request.user.SignInRequest
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
 import io.mockk.every
